+++ components/cdc/tests/failpoints/test_resolve.rs
@@ -1,4 +1,1 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
-mod test_observe;
-mod test_register;
-mod test_resolve;
